Fix compilation errors and warnings for C++20 compatibility
- Fix fatal error: use `rhs.originalParameters` instead of `rhs.parameters_` in RefinementParameters copy constructor (multiGridParameterManager.hh)
- Fix C++20 template-id warnings: remove `<T>` from constructor names in Xinterpolation2D, Yinterpolation2D, Zinterpolation2D (gridRefinementFunctional3D.hh)
- Fix sign-compare warning: cast `count` to `size_t` for comparison with `del_iX.size()` (complab3d_processors_part2.hh)
- Remove unused variables: `tortuosity_factor`, `safety_factor`, `estimated_iterations`, `bMin`, `bAvg` (complab.cpp)

+++ src/complab3d_processors_part2.hh
@@ -552,7 +552,7 @@ public:
                                 T nbrB = lattices[bMtLloc]->get(iXb+delx,iYb+dely,iZb+delz).computeDensity();
                                 if ( nbrAge >= age && (nbrB-Bmax)>-COMPLAB_THRD) { ++count; }
                             }
-                            if (count == del_iX.size()) { newAge = (T) (age+1); }
+                            if ((size_t)count == del_iX.size()) { newAge = (T) (age+1); }
                             else { updateflag = 0; }
                         }
                         if (updateflag == 1) {
